OJ-2983: Move common functionality to utils
Move getParameters and signer logic into util folder
