+++ test-resources/headless-core-stub/lambdas/callback/src/services/configuration-helper.ts
@@ -1,4 +1,4 @@
-import { getParametersValues } from "./get-parameters";
+import { getParametersValues } from "../../../utils/parameter/get-parameters";
 
 const commonParameterPrefix = process.env.AWS_STACK_NAME || "common-cri-api";
 const testResourcesParameterPrefix = process.env.TEST_RESOURCES_STACK_NAME || "test-resources";
