+++ test-resources/headless-core-stub/lambdas/utils/crypto/signer.ts
