+++ test-resources/headless-core-stub/lambdas/utils/parameter/get-parameters.ts
